Adding params to protect admin interface fixes #231

+++ lib/VCE/Services/User.pm
@@ -59,38 +59,41 @@ sub _register_methods{
     my $self = shift;
     my $dispatcher = shift;
 
-    my $method = GRNOC::WebService::Method->new( name => 'get_users',
-						 description => 'get a list of users',
-						 callback => sub { return $self->get_users(@_); }
+    my $method = GRNOC::WebService::Method->new(
+        name => 'get_users',
+        description => 'get a list of users',
+        callback => sub { return $self->get_users(@_); }
 	);
-    
     $method->add_input_parameter(
-	required => 0,
-	name => 'username',
-	pattern => $GRNOC::WebService::Regex::NAME_ID,
-	description => "username of the user");
-    
-    
+        required => 1,
+        name => 'workgroup',
+        pattern => $GRNOC::WebService::Regex::NAME_ID,
+        description => "name of current users workgroup"
+	);
+    $method->add_input_parameter(
+        required => 0,
+        name => 'username',
+        pattern => $GRNOC::WebService::Regex::NAME_ID,
+        description => "username of the user"
+    );
     $method->add_input_parameter(
         required => 0,
         name => 'email',
         pattern => $GRNOC::WebService::Regex::TEXT,
-        description => "email of the user");
-
-
+        description => "email of the user"
+    );
     $method->add_input_parameter(
         required => 0,
         name => 'fullname',
         pattern => $GRNOC::WebService::Regex::TEXT,
-        description => "fullname of the user");
-
-
+        description => "fullname of the user"
+    );
     $method->add_input_parameter(
         required => 0,
         name => 'user_id',
         pattern => $GRNOC::WebService::Regex::NUMBER_ID,
-        description => "user_id of the user to find");
-
+        description => "user_id of the user to find"
+    );
     $dispatcher->register_method($method);
 
     $method = GRNOC::WebService::Method->new(
@@ -100,69 +103,55 @@ sub _register_methods{
 	);
     $dispatcher->register_method($method);
 
-    $method = GRNOC::WebService::Method->new( name => 'add_user',
-					      description => 'get a list of commands and the details of those commands',
-					      callback => sub { return $self->add_user(@_); }
-        );
-
+    $method = GRNOC::WebService::Method->new(
+        name => 'modify_user',
+        description => 'Modifies and existing user',
+        callback => sub { return $self->modify_user(@_); }
+	);
     $method->add_input_parameter(
         required => 1,
-        name => 'username',
+        name => 'workgroup',
         pattern => $GRNOC::WebService::Regex::NAME_ID,
-        description => "username of the new user");
-
-    $method->add_input_parameter(
-	required => 0,
-	name => 'email',
-	pattern => $GRNOC::WebService::Regex::TEXT,
-	description => "Email address of the new user");
-
-    $method->add_input_parameter(
-	required => 0,
-	name => 'fullname',
-	pattern => $GRNOC::WebService::Regex::TEXT,
-	description => "Full Name of the new user");
-
-    $dispatcher->register_method($method);
-    
-    $method = GRNOC::WebService::Method->new( name => 'modify_user',
-                                              description => 'Modifies and existing user',
-					      callback => sub { return $self->modify_user(@_); }
+        description => "name of current users workgroup"
 	);
-
     $method->add_input_parameter(
         required => 0,
         name => 'email',
         pattern => $GRNOC::WebService::Regex::TEXT,
-        description => "Email address of the new user");
-
+        description => "Email address of the new user"
+    );
     $method->add_input_parameter(
         required => 0,
         name => 'fullname',
         pattern => $GRNOC::WebService::Regex::TEXT,
-        description => "Full Name of the new user");
-
-    $method->add_input_parameter( 
-	required => 1,
-	name => 'user_id',
-	pattern => $GRNOC::WebService::Regex::NUMBER_ID,
-	description => "User ID of the user to modify");
-
-    $dispatcher->register_method($method);
-
-    $method = GRNOC::WebService::Method->new( name => 'delete_user',
-                                              description => 'deletes and existing user',
-                                              callback => sub { return $self->delete_user(@_); }
-        );
-
+        description => "Full Name of the new user"
+    );
     $method->add_input_parameter(
         required => 1,
         name => 'user_id',
         pattern => $GRNOC::WebService::Regex::NUMBER_ID,
-	description => "the id of the user to delete");
-
+        description => "User ID of the user to modify"
+    );
     $dispatcher->register_method($method);
 
+    $method = GRNOC::WebService::Method->new(
+        name => 'delete_user',
+        description => 'deletes and existing user',
+        callback => sub { return $self->delete_user(@_); }
+    );
+    $method->add_input_parameter(
+        required => 1,
+        name => 'workgroup',
+        pattern => $GRNOC::WebService::Regex::NAME_ID,
+        description => "name of current users workgroup"
+	);
+    $method->add_input_parameter(
+        required => 1,
+        name => 'user_id',
+        pattern => $GRNOC::WebService::Regex::NUMBER_ID,
+        description => "the id of the user to delete"
+    );
+    $dispatcher->register_method($method);
 }
 
 =head2 get_current
@@ -187,30 +176,27 @@ sub get_users{
     my $m_ref = shift;
     my $p_ref = shift;
 
+    my $user = $ENV{'REMOTE_USER'};
+    my $workgroup = $p_ref->{'workgroup'}{'value'};
+
+    if (!$self->vce->access->user_in_workgroup(username => $user, workgroup => $workgroup )) {
+        $m_ref->set_error("User $user not in workgroup $workgroup.");
+        return;
+    }
+    my $is_admin = $self->vce->access->get_admin_workgroup()->{name} eq $workgroup ? 1 : 0;
+    if (!$is_admin) {
+        $m_ref->set_error("Workgroup $workgroup is not authorized to get users.");
+        return;
+    }
+
     my %args;
-    
+
     foreach my $key (keys (%{$p_ref})){
 	$args{$key} = $p_ref->{$key}{'value'};
     }
-    
+
     my $res = $self->db->get_users( %args );
     return {results => $res};
-}
-
-=head2 add_user
-
-=cut
-sub add_user{
-    my $self = shift;
-    my $m_ref = shift;
-    my $p_ref = shift;
-
-    my $username = $p_ref->{'username'}{'value'};
-    my $email = $p_ref->{'email'}{'value'};
-    my $fullname = $p_ref->{'fullname'}{'value'};
-
-    my $res = $self->db->add_user( $username, $email, $fullname);
-    return {results => [{id => $res}]};
 }
 
 =head2 modify_user
@@ -226,10 +212,25 @@ sub modify_user{
     my $fullname = $p_ref->{'fullname'}{'value'};
     my $email = $p_ref->{'email'}{'value'};
 
-    my $res = $self->db->modify_user(user_id => $user_id, 
-				     username => $username,
-				     fullname => $fullname,
-				     email => $email);
+    my $user = $ENV{'REMOTE_USER'};
+    my $workgroup = $p_ref->{'workgroup'}{'value'};
+
+    if (!$self->vce->access->user_in_workgroup(username => $user, workgroup => $workgroup )) {
+        $m_ref->set_error("User $user not in workgroup $workgroup.");
+        return;
+    }
+    my $is_admin = $self->vce->access->get_admin_workgroup()->{name} eq $workgroup ? 1 : 0;
+    if (!$is_admin) {
+        $m_ref->set_error("Workgroup $workgroup is not authorized to modify users.");
+        return;
+    }
+
+    my $res = $self->db->modify_user(
+        user_id => $user_id,
+        username => $username,
+        fullname => $fullname,
+        email => $email
+    );
     if ($res eq "0E0") {
         $m_ref->set_error("Update failed for user: $user_id");
         return;
@@ -244,11 +245,23 @@ sub delete_user{
     my $self = shift;
     my $m_ref = shift;
     my $p_ref = shift;
-    
-    my $user_id = $p_ref->{'user_id'}{'value'};
-    
-    my $res = $self->db->delete_user( $user_id );
 
+    my $user = $ENV{'REMOTE_USER'};
+    my $workgroup = $p_ref->{'workgroup'}{'value'};
+
+    if (!$self->vce->access->user_in_workgroup(username => $user, workgroup => $workgroup )) {
+        $m_ref->set_error("User $user not in workgroup $workgroup.");
+        return;
+    }
+    my $is_admin = $self->vce->access->get_admin_workgroup()->{name} eq $workgroup ? 1 : 0;
+    if (!$is_admin) {
+        $m_ref->set_error("Workgroup $workgroup is not authorized to delete users.");
+        return;
+    }
+
+    my $user_id = $p_ref->{'user_id'}{'value'};
+
+    my $res = $self->db->delete_user($user_id);
     if ($res eq "0E0") {
         $m_ref->set_error("Delete failed for user: $user_id");
         return;
